Add endpoints for file creation times and current time; refactor static file serving
