+++ AccountManager/frontend/client/.env
@@ -1,2 +1,2 @@
-VITE_BASE_URL= https://app.technests.ai/
-# VITE_BASE_URL = http://localhost:3000/
+# VITE_BASE_URL= https://app.technests.ai/
+VITE_BASE_URL = http://localhost:3000/
